refactor: APIProtocol, HttpMethod 별도 파일로 분리 #2
