+++ OpenMarket/OpenMarket/Protocol/APIProtocol.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+protocol APIProtocol {
+    var url: URL? { get }
+    var method: HttpMethod { get }
+}
+
+enum HttpMethod: CustomStringConvertible {
+    case get
+    
+    var description: String {
+        switch self {
+        case .get:
+            return "GET"
+        }
+    }
+}
